+++ xtask/src/node_versions.rs
@@ -18,9 +18,9 @@ pub fn build_node_versions() -> Result<()> {
     let versions = releases.into_iter().map(|release| {
         let version = release.version.split('.').collect::<Vec<_>>();
         assert_eq!(version.len(), 3);
-        let major: u32 = version[0].parse().unwrap();
-        let minor: u32 = version[1].parse().unwrap();
-        let patch: u32 = version[2].parse().unwrap();
+        let major: u8 = version[0].parse().unwrap();
+        let minor: u8 = version[1].parse().unwrap();
+        let patch: u8 = version[2].parse().unwrap();
         quote! {
             Version(#major, #minor, #patch)
         }
